refa: rename extconfig to config
As the config package doesn't implement any extension point I'd suggest
that we rename it to config.
This also reads a bit more nicely when using it and also alignes the
scaffold with the other extensions already implemented.
